Fixed bug in 'drop table' Portability fixes
sql-bench/bench-init.pl.sh:
  Change mode to -rw-rw-r--
sql-bench/compare-results.sh:
  Change mode to -rw-rw-r--
sql-bench/run-all-tests.sh:
  Change mode to -rw-rw-r--
sql-bench/test-ATIS.sh:
  Change mode to -rw-rw-r--
sql-bench/test-alter-table.sh:
  Change mode to -rw-rw-r--
sql-bench/test-big-tables.sh:
  Change mode to -rw-rw-r--
sql-bench/test-connect.sh:
  Change mode to -rw-rw-r--
sql-bench/test-create.sh:
  Change mode to -rw-rw-r--
sql-bench/test-select.sh:
  Change mode to -rw-rw-r--
sql-bench/test-wisconsin.sh:
  Change mode to -rw-rw-r--
Docs/manual.texi:
  Added documentation for perror
myisam/mi_check.c:
  Portability fix
mysql-test/chew_on_this/select.res:
  ***MISSING WEAVE***
scripts/mysqlhotcopy.sh:
  Added --noindices
sql-bench/Results/ATIS-mysql-Linux_2.2.13_SMP_alpha:
  New benchmark results
sql-bench/Results/RUN-mysql-Linux_2.2.13_SMP_alpha:
  New benchmark results
sql-bench/Results/alter-table-mysql-Linux_2.2.13_SMP_alpha:
  New benchmark results
sql-bench/Results/big-tables-mysql-Linux_2.2.13_SMP_alpha:
  New benchmark results
sql-bench/Results/connect-mysql-Linux_2.2.13_SMP_alpha:
  New benchmark results
sql-bench/Results/create-mysql-Linux_2.2.13_SMP_alpha:
  New benchmark results
sql-bench/Results/insert-mysql-Linux_2.2.13_SMP_alpha:
  New benchmark results
sql-bench/Results/select-mysql-Linux_2.2.13_SMP_alpha:
  New benchmark results
sql-bench/Results/wisconsin-mysql-Linux_2.2.13_SMP_alpha:
  New benchmark results
sql-bench/copy-db.sh:
  Changed multi_value_insert -> insert_multi_value
sql-bench/crash-me.sh:
  Changed multi_value_insert -> insert_multi_value
sql-bench/limits/Adabas.cfg:
  Changed multi_value_insert -> insert_multi_value
sql-bench/limits/access.cfg:
  Changed multi_value_insert -> insert_multi_value
sql-bench/limits/access_odbc.cfg:
  Changed multi_value_insert -> insert_multi_value
sql-bench/limits/frontbase.cfg:
  Changed multi_value_insert -> insert_multi_value
sql-bench/limits/interbase.cfg:
  Changed multi_value_insert -> insert_multi_value
sql-bench/limits/mimer.cfg:
  Changed multi_value_insert -> insert_multi_value
sql-bench/limits/ms-sql.cfg:
  Changed multi_value_insert -> insert_multi_value
sql-bench/limits/msql.cfg:
  Changed multi_value_insert -> insert_multi_value
sql-bench/limits/mysql-3.22.cfg:
  Changed multi_value_insert -> insert_multi_value
sql-bench/limits/mysql-3.23.cfg:
  Changed multi_value_insert -> insert_multi_value
sql-bench/limits/mysql.cfg:
  Changed multi_value_insert -> insert_multi_value
sql-bench/limits/oracle.cfg:
  Changed multi_value_insert -> insert_multi_value
sql-bench/limits/pg.cfg:
  Changed multi_value_insert -> insert_multi_value
sql-bench/print-limit-table:
  Better groups
sql-bench/server-cfg.sh:
  Changed multi_value_insert -> insert_multi_value
sql-bench/test-insert.sh:
  Changed multi_value_insert -> insert_multi_value
sql/ha_myisam.cc:
  Added warning of number of rows changed in optimize
sql/mysqlbinlog.cc:
  Portability fix
sql/mysqld.cc:
  Changed mysqld to take less stack and safer shutdown
sql/share/swedish/errmsg.OLD:
  update
sql/sql_base.cc:
  Fixed 'drop table' bug
sql/sql_parse.cc:
  Fixed bug in 'TRUNCATE'
sql/sql_show.cc:
  Don't lock keycache when you do 'show variables'

+++ myisam/mi_check.c
@@ -108,7 +108,7 @@ int chk_status(MI_CHECK *param, register MI_INFO *info)
   else if (mi_is_crashed(info))
     mi_check_print_warning(param,
 			   "Table is marked as crashed");
-  if (share->state.open_count != (info->s->global_changed ? 1 : 0))
+  if (share->state.open_count != (uint) (info->s->global_changed ? 1 : 0))
   {
     mi_check_print_warning(param,
 			   "%d clients is using or hasn't closed the table properly",
